[RFR-520] Move pre-oauth classes to v2 package
