+++ api/src/test/java/de/cyface/api/v2/AuthorizationTest.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with the Cyface API Library. If not, see <http://www.gnu.org/licenses/>.
  */
-package de.cyface.api;
+package de.cyface.api.v2;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.hasItems;
@@ -30,7 +30,12 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import de.cyface.api.AuthenticatedEndpointConfig;
+import de.cyface.api.DatabaseConstants;
+import de.cyface.api.EndpointConfig;
+import de.cyface.api.PauseAndResumeAfterBodyParsing;
 import org.apache.commons.lang3.Validate;
+import org.hamcrest.MatcherAssert;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -39,8 +44,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import de.cyface.api.model.Role;
-import de.cyface.api.model.User;
+import de.cyface.api.v2.model.Role;
+import de.cyface.api.v2.model.User;
 import de.cyface.apitestutils.TestMongoDatabase;
 import de.cyface.apitestutils.fixture.user.DirectTestUser;
 import de.cyface.apitestutils.fixture.user.TestUser;
@@ -139,7 +144,7 @@ public class AuthorizationTest {
                         // Assert
                         final var expectedUsernames = env.getExpectedUsernames();
                         assertThat(users, hasSize(expectedUsernames.size()));
-                        assertThat(users.stream().map(User::getName).collect(Collectors.toList()),
+                        MatcherAssert.assertThat(users.stream().map(User::getName).collect(Collectors.toList()),
                                 hasItems(expectedUsernames.toArray(new String[0])));
                         testContext.completeNow();
                     }));
@@ -180,7 +185,7 @@ public class AuthorizationTest {
                         final var expectedUsernames = List
                                 .of(new String[] {groupUser.getUsername(), groupManager.getUsername()});
                         assertThat(users, hasSize(expectedUsernames.size()));
-                        assertThat(users.stream().map(User::getName).collect(Collectors.toList()),
+                        MatcherAssert.assertThat(users.stream().map(User::getName).collect(Collectors.toList()),
                                 hasItems(expectedUsernames.toArray(new String[0])));
                         testContext.completeNow();
                     }));
